Rename shared url file
